update operator edit page
